Work on issue observable
