+++ src/epics/index.ts
@@ -1,7 +1,9 @@
 import userEpics from './user';
 import repositoryEpics from './repositories';
+import issueEpics from './issues';
 
 export default [
   userEpics,
+  issueEpics,
   repositoryEpics
 ];
