+++ src/epics/repositories.ts
@@ -37,42 +37,3 @@ const fetchTotalReposLengthEpic = (action$) => (
 );
 
 export default combineEpics(fetchReposEpic, fetchTotalReposLengthEpic);
-
-// export const fetchTotalReposLength = (username) => {
-//   return dispatch => {
-//     return get({
-//       endpoint: `users/${username}/starred`,
-//       params: { per_page: 1 },
-//       resHeader: true
-//     })
-//     // .then(headers => {
-//     //   const reg = /rel="next", <.*&page=(\d+)>; rel="last"/i;
-//     //   const len = parseInt(headers.get('link').match(reg)[1], 10);
-//     //   return append('starred', len)(dispatch);
-//     // });
-//   };
-// };
-
-// export const getRepos = (username, page) => {
-//   return dispatch => {
-//     return get({
-//       endpoint: `users/${username}/starred`,
-//       params: { page }
-//     })
-//     // .then((repos: List<Repository>) => {
-//     //   add(repos)(dispatch);
-//     //   return repos;
-//     // })
-//     // .then((repos: List<any>) => {
-//     //   const proms: Array<Promise<any>> = repos.map(repo =>
-//     //     getIssuesReq(repo.get('full_name'), repo.get('id'))(dispatch)
-//     //   ).toArray();
-
-//     //   return Promise.all(proms);
-//     // })
-//     // .then((pArr: Array<any>) => {
-//     //   const issues: List<any> = List(pArr).flatten(1).toList();
-//     //   return serializeIssues(issues)(dispatch);
-//     // });
-//   };
-// };
